backend overhaul with services and controller datacontracts

+++ maxhanna.Server/Controllers/CalendarController.cs
@@ -18,10 +18,10 @@ namespace maxhanna.Server.Controllers
         }
 
 
-        [HttpGet(Name = "GetCalendar")]
-        public async Task<IActionResult> Get([FromQuery] DateTime startDate, [FromQuery] DateTime endDate)
+        [HttpPost(Name = "GetCalendar")]
+        public async Task<IActionResult> Get([FromBody] User user, [FromQuery] DateTime startDate, [FromQuery] DateTime endDate)
         {
-            _logger.LogInformation($"GET /Calendar (startDate : {startDate}) (endDate : {endDate})");
+            _logger.LogInformation($"GET /Calendar (startDate : {startDate}, endDate : {endDate}, user: {user.Id})");
             if (startDate > endDate)
             {
                 _logger.LogError("An error occurred while fetching calendar entries. StartDate > EndDate");
@@ -36,8 +36,8 @@ namespace maxhanna.Server.Controllers
                     await conn.OpenAsync();
 
                     string sql = 
-                        "SELECT Id, Type, Note, Date FROM maxhanna.calendar " +
-                        "WHERE " +
+                        "SELECT Id, Type, Note, Date, Ownership FROM maxhanna.calendar " +
+                        "WHERE Ownership = @Owner AND " +
                             "(" +
                                 "(Date BETWEEN (@StartDate - interval 1 day) AND @EndDate) " +
                                 "OR " +
@@ -47,6 +47,7 @@ namespace maxhanna.Server.Controllers
                             ") ";
                     using (var cmd = new MySqlCommand(sql, conn))
                     {
+                        cmd.Parameters.AddWithValue("@Owner", user.Id);
                         cmd.Parameters.AddWithValue("@StartDate", startDate);
                         cmd.Parameters.AddWithValue("@EndDate", endDate);
 
@@ -54,7 +55,7 @@ namespace maxhanna.Server.Controllers
                         {
                             while (await rdr.ReadAsync())
                             {
-                                entries.Add(new CalendarEntry(rdr.GetInt32(0), rdr.GetString(1), rdr.GetString(2), rdr.GetDateTime(3)));
+                                entries.Add(new CalendarEntry(rdr.GetInt32(0), rdr.GetString(1), rdr.GetString(2), rdr.GetDateTime(3), rdr.GetString(4)));
                             }
                         }
                     }
@@ -69,23 +70,24 @@ namespace maxhanna.Server.Controllers
             }
         }
 
-        [HttpPost(Name = "CreateCalendarEntry")]
-        public async Task<IActionResult> Post([FromBody] CalendarEntry model)
+        [HttpPost("/Calendar/Create", Name = "CreateCalendarEntry")]
+        public async Task<IActionResult> Post([FromBody] CreateCalendarEntry req)
         {
             _logger.LogInformation("POST /Calendar");
-            _logger.LogInformation($"Type : {model.Type} Note: {model.Note} Date: {model.Date}");
+            _logger.LogInformation($"Type : {req.calendarEntry.Type} Note: {req.calendarEntry.Note} Date: {req.calendarEntry.Date}, User: {req.user.Id}");
             MySqlConnection conn = new MySqlConnection(_config.GetValue<string>("ConnectionStrings:maxhanna"));
             try
             {
                 conn.Open();
 
                 // Assuming CalendarEntryModel has properties for Type, Note, and Date
-                string sql = "INSERT INTO maxhanna.calendar (Type, Note, Date) VALUES (@Type, @Note, @Date)";
+                string sql = "INSERT INTO maxhanna.calendar (Type, Note, Date, Ownership) VALUES (@Type, @Note, @Date, @Owner)";
                 MySqlCommand cmd = new MySqlCommand(sql, conn);
-                cmd.Parameters.AddWithValue("@Type", model.Type);
-                cmd.Parameters.AddWithValue("@Note", model.Note);
-                cmd.Parameters.AddWithValue("@Date", model.Date);
-                _logger.LogInformation("note : " + model.Note);
+                cmd.Parameters.AddWithValue("@Type", req.calendarEntry.Type);
+                cmd.Parameters.AddWithValue("@Note", req.calendarEntry.Note);
+                cmd.Parameters.AddWithValue("@Date", req.calendarEntry.Date);
+                cmd.Parameters.AddWithValue("@Owner", req.user.Id);
+                _logger.LogInformation("note : " + req.calendarEntry.Note);
                 if (await cmd.ExecuteNonQueryAsync() > 0)
                 {
                     _logger.LogInformation("Returned OK");
@@ -109,17 +111,18 @@ namespace maxhanna.Server.Controllers
         }
 
         [HttpDelete("{id}", Name = "DeleteCalendarEntry")]
-        public async Task<IActionResult> Delete(int id)
+        public async Task<IActionResult> Delete([FromBody] User user, int id)
         {
-            _logger.LogInformation($"Inside DELETE() of CalendarController for ID: {id}");
+            _logger.LogInformation($"Inside DELETE() of CalendarController for User: {user.Id} , CalendarEntryID: {id}");
             MySqlConnection conn = new MySqlConnection(_config.GetValue<string>("ConnectionStrings:maxhanna"));
             try
             {
                 conn.Open();
 
-                string sql = "DELETE FROM maxhanna.calendar WHERE ID = @Id";
+                string sql = "DELETE FROM maxhanna.calendar WHERE ID = @Id AND Ownership = @Owner";
                 MySqlCommand cmd = new MySqlCommand(sql, conn);
                 cmd.Parameters.AddWithValue("@Id", id);
+                cmd.Parameters.AddWithValue("@Owner", user.Id);
                 int rowsAffected = await cmd.ExecuteNonQueryAsync();
 
                 if (rowsAffected > 0)
