+++ maxhanna.Server/Controllers/ContactController.cs
@@ -17,12 +17,12 @@ namespace maxhanna.Server.Controllers
             _config = config;
         }
 
-        [HttpGet(Name = "GetContacts")]
-        public async Task<IActionResult> GetContacts()
+        [HttpPost(Name = "GetContacts")]
+        public async Task<IActionResult> GetContacts([FromBody] User user)
         {
-            _logger.LogInformation("GET /Contact");
+            _logger.LogInformation($"GET /Contact (User : {user.Id}) ");
 
-            string sql = "SELECT id, name, phone, birthday, notes, email FROM contacts";
+            string sql = "SELECT id, name, phone, birthday, notes, email, ownership FROM contacts";
 
             try
             {
@@ -45,7 +45,8 @@ namespace maxhanna.Server.Controllers
                                     Phone = rdr.IsDBNull(2) ? null : rdr.GetString(2),
                                     Birthday = rdr.IsDBNull(3) ? null : rdr.GetDateTime(3),
                                     Notes = rdr.IsDBNull(4) ? null : rdr.GetString(4),
-                                    Email = rdr.IsDBNull(5) ? null : rdr.GetString(5)
+                                    Email = rdr.IsDBNull(5) ? null : rdr.GetString(5),
+                                    Ownership = rdr.IsDBNull(6) ? null : rdr.GetString(6)
                                 });
                             }
 
@@ -61,22 +62,23 @@ namespace maxhanna.Server.Controllers
             }
         }
 
-        [HttpPost(Name = "CreateContact")]
-        public async Task<IActionResult> CreateContact([FromBody] Contact model)
+        [HttpPost("/Contact/Create", Name = "CreateContact")]
+        public async Task<IActionResult> CreateContact([FromBody] CreateContact req)
         {
-            _logger.LogInformation("POST /Contact");
+            _logger.LogInformation($"POST /Contact/Create (User: {req.user.Id})");
             MySqlConnection conn = new MySqlConnection(_config.GetConnectionString("maxhanna"));
 
             try
             {
                 await conn.OpenAsync();
-                string sql = "INSERT INTO contacts (name, phone, birthday, notes, email) VALUES (@Name, @Phone, @Birthday, @Notes, @Email)";
+                string sql = "INSERT INTO contacts (name, phone, birthday, notes, email, ownership) VALUES (@Name, @Phone, @Birthday, @Notes, @Email, @Owner)";
                 MySqlCommand cmd = new MySqlCommand(sql, conn);
-                cmd.Parameters.AddWithValue("@Name", model.Name);
-                cmd.Parameters.AddWithValue("@Phone", model.Phone ?? (object)DBNull.Value);
-                cmd.Parameters.AddWithValue("@Birthday", model.Birthday ?? (object)DBNull.Value);
-                cmd.Parameters.AddWithValue("@Notes", model.Notes ?? (object)DBNull.Value);
-                cmd.Parameters.AddWithValue("@Email", model.Email ?? (object)DBNull.Value);
+                cmd.Parameters.AddWithValue("@Name", req.contact.Name);
+                cmd.Parameters.AddWithValue("@Phone", req.contact.Phone ?? (object)DBNull.Value);
+                cmd.Parameters.AddWithValue("@Birthday", req.contact.Birthday ?? (object)DBNull.Value);
+                cmd.Parameters.AddWithValue("@Notes", req.contact.Notes ?? (object)DBNull.Value);
+                cmd.Parameters.AddWithValue("@Email", req.contact.Email ?? (object)DBNull.Value);
+                cmd.Parameters.AddWithValue("@Owner", req.user.Id);
                 await cmd.ExecuteNonQueryAsync();
 
                 _logger.LogInformation("Returned OK");
@@ -94,22 +96,23 @@ namespace maxhanna.Server.Controllers
         }
 
         [HttpPut("/Contact/{id}", Name = "UpdateContact")]
-        public async Task<IActionResult> UpdateContact(int id, [FromBody] Contact model)
+        public async Task<IActionResult> UpdateContact(int id, [FromBody] CreateContact req)
         {
-            _logger.LogInformation($"PUT /Contact/{id}");
+            _logger.LogInformation($"PUT /Contact/{id} (User: {req.user.Id})");
             MySqlConnection conn = new MySqlConnection(_config.GetConnectionString("maxhanna"));
 
             try
             {
                 await conn.OpenAsync();
-                string sql = "UPDATE contacts SET name = @Name, phone = @Phone, birthday = @Birthday, notes = @Notes, email = @Email WHERE id = @Id";
+                string sql = "UPDATE contacts SET name = @Name, phone = @Phone, birthday = @Birthday, notes = @Notes, email = @Email WHERE id = @Id AND ownership = @Owner;";
                 MySqlCommand cmd = new MySqlCommand(sql, conn);
-                cmd.Parameters.AddWithValue("@Name", model.Name);
-                cmd.Parameters.AddWithValue("@Phone", model.Phone ?? (object)DBNull.Value);
-                cmd.Parameters.AddWithValue("@Birthday", model.Birthday ?? (object)DBNull.Value);
-                cmd.Parameters.AddWithValue("@Notes", model.Notes ?? (object)DBNull.Value);
-                cmd.Parameters.AddWithValue("@Email", model.Email ?? (object)DBNull.Value);
+                cmd.Parameters.AddWithValue("@Name", req.contact.Name);
+                cmd.Parameters.AddWithValue("@Phone", req.contact.Phone ?? (object)DBNull.Value);
+                cmd.Parameters.AddWithValue("@Birthday", req.contact.Birthday ?? (object)DBNull.Value);
+                cmd.Parameters.AddWithValue("@Notes", req.contact.Notes ?? (object)DBNull.Value);
+                cmd.Parameters.AddWithValue("@Email", req.contact.Email ?? (object)DBNull.Value);
                 cmd.Parameters.AddWithValue("@Id", id);
+                cmd.Parameters.AddWithValue("@Owner", req.user.Id);
                 int rowsAffected = await cmd.ExecuteNonQueryAsync();
 
                 if (rowsAffected > 0)
@@ -133,18 +136,21 @@ namespace maxhanna.Server.Controllers
         }
 
         [HttpDelete("{id}")]
-        public async Task<IActionResult> Delete(int id)
+        public async Task<IActionResult> Delete([FromBody] User user, int id)
         {
+            _logger.LogInformation($"DELETE /Contact/{id} (User: {user.Id})");
+
             try
             {
                 using (var connection = new MySqlConnection(_config.GetConnectionString("maxhanna")))
                 {
                     await connection.OpenAsync();
 
-                    string query = "DELETE FROM contacts WHERE id = @Id";
+                    string query = "DELETE FROM contacts WHERE id = @Id AND ownership = @Owner";
                     using (var command = new MySqlCommand(query, connection))
                     {
                         command.Parameters.AddWithValue("@Id", id);
+                        command.Parameters.AddWithValue("@Owner", user.Id);
 
                         int rowsAffected = await command.ExecuteNonQueryAsync();
                         if (rowsAffected > 0)
